Add support for custom target to scroll to.
The targetElement can be specified in the options object the same way as element.
The solution basically just replaces the difference param of the easing function with endPosition-startPosition.
(A util function is added to calculate the target DOM element's offset)
A default value of 0 ensures backwards compatibility with the old functionality.